+++ resources/leiningen/new/lein_compojure/app_test.clj
@@ -1,13 +1,23 @@
 (ns {{name}}.app-test
   (:require [clojure.test :refer [deftest is testing]]
+            [environ.core :refer [env]]
+            [ring-basic-auth-test :as basic-auth]
             [ring.mock.request :as mock]
             [{{name}}.app :refer [handler]]))
 
+(def with-auth
+  (basic-auth/authorize (:http-basic-auth-username env)
+                        (:http-basic-auth-password env)))
+
 (deftest site-routes-test
   (testing "main route"
-    (let [response (handler (mock/request :get "/"))]
+    (let [response (handler (with-auth (mock/request :get "/")))]
       (is (= 200 (:status response) 200))))
 
   (testing "not-found route"
-    (let [response (handler (mock/request :get "/invalid"))]
-      (is (= 404 (:status response) 404)))))
+    (let [response (handler (with-auth (mock/request :get "/invalid")))]
+      (is (= 404 (:status response) 404))))
+
+  (testing "main route when not authenticated"
+    (let [response (handler (mock/request :get "/"))]
+      (is (= "" (:status response))))))